Fix MissingInformation class error
On sway, each workspace has a non-floating pseudo root container which
hold all the other containers. This container is pretty basic and does
not have any name.

When looking for window name, we loop through those containers and add
them in the vector of valid windows. Later in the code, we retrieve the
name of all the windows in the vector, and fall in the caveat of those
empty-named containers.

Add a workaround not to add the 'non-floating' root containers.

Signed-off-by: Raphael Gallais-Pou <rgallaispou@gmail.com>